src/processor_function/requirements.txt
functions-framework==3.*
google-cloud-firestore
google-api-python-client